+++ src/Microsoft/LoggerExtensions.cs
@@ -373,6 +373,7 @@ public static partial class LoggerExtensions
 	/// <param name="logger"> The extended <see cref="ILogger"/>. </param>
 	/// <param name="scope"> The scope to apply. </param>
 	/// <returns> The logging scope. </returns>
+	/// <remarks> It may be better to get the groups of a logger with the <see cref="AsGroup{TIdentifier}"/> extension method and then applying the scope with one of the group methods like <see cref="ILoggerGroup.CreateScope(LogScope)"/>. </remarks>
 	public static IDisposable CreateScope<TIdentifier>(this ILogger logger, LogScope<TIdentifier>? scope)
 		where TIdentifier : notnull
 		=> scope == null ? NoDisposable.Instance : logger.AsGroup(scope.Identifier).CreateScope(scope);
@@ -548,6 +549,112 @@ public static partial class LoggerExtensions
 			cleanCallerArgument
 		);
 		return logger.BeginScope(scopes);
+	}
+
+#endif
+
+	/// <summary>
+	/// Creates a new logging <paramref name="scope"/> that is not removable.
+	/// </summary>
+	/// <param name="logger"> The extended <see cref="ILogger"/>. </param>
+	/// <param name="scope"> The scope to apply. </param>
+	/// <returns> The same <see cref="ILogger"/> instance for chaining. </returns>
+	public static ILogger PinScope(this ILogger logger, LogScope? scope)
+	{
+		if (scope is not null)
+			logger.BeginScope((IDictionary<string, object?>) scope);
+		return logger;
+	}
+
+	/// <summary>
+	/// Creates a new logging scope with named values that is not removable.
+	/// </summary>
+	/// <param name="logger"> The extended <see cref="ILogger"/>. </param>
+	/// <param name="scopedValues"> Collection of named values. </param>
+	/// <returns> The same <see cref="ILogger"/> instance for chaining. </returns>
+	public static ILogger PinScope(this ILogger logger, params (string Identifier, object? Value)[] scopedValues)
+	{
+		var scopes = LogScopeBuilder.BuildScopeDictionary(scopedValues);
+		logger.BeginScope(scopes);
+		return logger;
+	}
+
+	/// <summary>
+	/// Creates a new logging scope with named values extracted from the given <see cref="Expression"/>s that is not removable.
+	/// </summary>
+	/// <param name="logger"> The extended <see cref="ILogger"/>. </param>
+	/// <param name="scopedValues"> The <see cref="Expression"/>s used to build the named values. </param>
+	/// <returns> The same <see cref="ILogger"/> instance for chaining. </returns>
+	public static ILogger PinScope(this ILogger logger, params Expression<Func<object>>[] scopedValues)
+	{
+		var scopes = LogScopeBuilder.BuildScopeDictionary(scopedValues);
+		logger.BeginScope(scopes);
+		return logger;
+	}
+
+#if NETCOREAPP3_0_OR_GREATER
+
+	/// <summary>
+	/// Creates a new logging scope with named values extracted from the given <see cref="Expression"/>s that is not removable.
+	/// </summary>
+	/// <param name="logger"> The extended <see cref="ILogger"/>. </param>
+	/// <param name="value1"> The value that will be added to the scope. </param>
+	/// <param name="name1"> The expression name of <paramref name="value1"/> obtained via 'System.Runtime.CompilerServices.CallerArgumentExpression'. </param>
+	/// <param name="value2"> See: <paramref name="name1"/>. </param>
+	/// <param name="name2"> See: <paramref name="value1"/>. </param>
+	/// <param name="value3"> See: <paramref name="name1"/>. </param>
+	/// <param name="name3"> See: <paramref name="value1"/>. </param>
+	/// <param name="value4"> See: <paramref name="name1"/>. </param>
+	/// <param name="name4"> See: <paramref name="value1"/>. </param>
+	/// <param name="value5"> See: <paramref name="name1"/>. </param>
+	/// <param name="name5"> See: <paramref name="value1"/>. </param>
+	/// <param name="value6"> See: <paramref name="name1"/>. </param>
+	/// <param name="name6"> See: <paramref name="value1"/>. </param>
+	/// <param name="value7"> See: <paramref name="name1"/>. </param>
+	/// <param name="name7"> See: <paramref name="value1"/>. </param>
+	/// <param name="value8"> See: <paramref name="name1"/>. </param>
+	/// <param name="name8"> See: <paramref name="value1"/>. </param>
+	/// <param name="value9"> See: <paramref name="name1"/>. </param>
+	/// <param name="name9"> See: <paramref name="value1"/>. </param>
+	/// <param name="value10"> See: <paramref name="name1"/>. </param>
+	/// <param name="name10"> See: <paramref name="value1"/>. </param>
+	/// <param name="cleanCallerArgument"> Should the caller argument parameter be cleaned (removes everything but the last section of a <b>dot</b> separated string). Default is <b>true</b>. </param>
+	/// <returns> The same <see cref="ILogger"/> instance for chaining. </returns>
+	/// <exception cref="ArgumentNullException"> Is thrown if any name could not be automatically obtained while its value is specified. </exception>
+	public static ILogger PinScope
+	(
+		this ILogger logger,
+		object? value1,
+		object? value2 = default,
+		object? value3 = default,
+		object? value4 = default,
+		object? value5 = default,
+		object? value6 = default,
+		object? value7 = default,
+		object? value8 = default,
+		object? value9 = default,
+		object? value10 = default,
+		[System.Runtime.CompilerServices.CallerArgumentExpression("value1")] string? name1 = default,
+		[System.Runtime.CompilerServices.CallerArgumentExpression("value2")] string? name2 = default,
+		[System.Runtime.CompilerServices.CallerArgumentExpression("value3")] string? name3 = default,
+		[System.Runtime.CompilerServices.CallerArgumentExpression("value4")] string? name4 = default,
+		[System.Runtime.CompilerServices.CallerArgumentExpression("value5")] string? name5 = default,
+		[System.Runtime.CompilerServices.CallerArgumentExpression("value6")] string? name6 = default,
+		[System.Runtime.CompilerServices.CallerArgumentExpression("value7")] string? name7 = default,
+		[System.Runtime.CompilerServices.CallerArgumentExpression("value8")] string? name8 = default,
+		[System.Runtime.CompilerServices.CallerArgumentExpression("value9")] string? name9 = default,
+		[System.Runtime.CompilerServices.CallerArgumentExpression("value10")] string? name10 = default,
+		bool cleanCallerArgument = true
+	)
+	{
+		var scopes = LogScopeBuilder.BuildScopeDictionary
+		(
+			value1, value2, value3, value4, value5, value6, value7, value8, value9, value10,
+			name1, name2, name3, name4, name5, name6, name7, name8, name9, name10,
+			cleanCallerArgument
+		);
+		logger.BeginScope(scopes);
+		return logger;
 	}
 
 #endif