+++ src/Microsoft/⬙/CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ___
 
+## 2.2.0
+
+:calendar: _2023-03-25_
+
+### Added
+
+- New `ILogger` extension methods `PinScope` that create log scopes that are not removable and return the same `ILogger` instance for chaining. This can be used when initially setting up logger instances.
+___
+
 ## 2.1.0
 
 :calendar: _2023-03-16_